Reduce scene1 bundle size by 1.2 KB
Three small, safe bundle-size reductions targeting scene1's eager entry chunk:

1. arc-rotate-controls: compress 16 addEventListener/removeEventListener calls into a [event, handler, opts] table driven by two short loops.

2. load-env: flatten polynomialToPreScaledHarmonics using 7 pre-folded bit-exact SH constants and a single 3-iteration loop. Removes per-call scale/add/sub helpers and intermediate Float32Array allocations.

3. math/mat4: split mat4Invert into its own mat4-invert.ts module so Rollup can place it in the gltf-animation/picking chunk instead of pinning it to the eager entry chunk via mat4.ts. 1.73 KB moves out of scene1's entry.

Results (raw bundle size):

  scene1:  90.4 -> 89.2 KB (-1.2)

  scene7:  101.5 -> 101.1 KB

  scene12: 95.8 -> 95.4 KB

  scene14: 90.6 -> 89.5 KB

  scene34: 85.8 -> 85.4 KB

All 68 parity tests pass; no ceiling or golden-reference changes.
